yolo/requirements.txt
# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.2.34  # https://ultralytics.com
Flask>=2.2.0,<2.3.0  # Flask 버전 설정

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0